+++ example/spec/helpers/granuleUtils.js
@@ -10,7 +10,6 @@ const path = require('path');
 const pWaitFor = require('p-wait-for');
 
 const { buildS3Uri } = require('@cumulus/aws-client/S3');
-const { Granule } = require('@cumulus/api/models');
 const { s3 } = require('@cumulus/aws-client/services');
 const { randomStringFromRegex } = require('@cumulus/common/test-utils');
 const {
@@ -145,29 +144,20 @@ async function setupTestGranuleForIngest(bucket, inputPayloadJson, granuleRegex,
 }
 
 const deleteGranules = async (prefix, granules) => {
-  process.env.GranulesTable = `${prefix}-GranulesTable`;
-  const granuleModel = new Granule();
-  return await Promise.all(
+  await Promise.all(
     granules.map(async (granule) => {
-      // Temporary fix to handle granules that are in a bad state
-      // and cannot be deleted via the API
-      try {
-        if (granule.published === true) {
-          return await removePublishedGranule({
-            prefix,
-            granuleId: granule.granuleId,
-          });
-        }
-        return await deleteGranule({
+    // Temporary fix to handle granules that are in a bad state
+    // and cannot be deleted via the API
+      if (granule.published === true) {
+        return await removePublishedGranule({
           prefix,
           granuleId: granule.granuleId,
         });
-      } catch (error) {
-        if (error.statusCode === 400 && JSON.parse(error.apiMessage).message.includes('validation errors')) { // TODO wat
-          return await granuleModel.delete({ granuleId: granule.granuleId });
-        }
-        throw error;
       }
+      return await deleteGranule({
+        prefix,
+        granuleId: granule.granuleId,
+      });
     })
   );
 };