+++ packages/api/app/routes.js
@@ -7,7 +7,6 @@ const { defaultErrorHandler } = require('./middleware');
 
 const collections = require('../endpoints/collections');
 const granules = require('../endpoints/granules');
-const granuleCsv = require('../endpoints/granule-csv');
 const providers = require('../endpoints/providers');
 const pdrs = require('../endpoints/pdrs');
 const rules = require('../endpoints/rules');
@@ -47,9 +46,6 @@ router.use('/collections', ensureAuthorized, collections.router);
 
 // granules endpoints
 router.use('/granules', ensureAuthorized, granules.router);
-
-// granule csv endpoints
-router.use('/granule-csv', ensureAuthorized, granuleCsv);
 
 // provider endpoints
 router.use('/providers', ensureAuthorized, providers.router);