+++ packages/api/models/index.js
@@ -3,7 +3,6 @@
 const AccessToken = require('./access-tokens');
 const AsyncOperation = require('./async-operation');
 const Manager = require('./base');
-const Granule = require('./granules');
 const ReconciliationReport = require('./reconciliation-reports');
 const Rule = require('./rules');
 const Execution = require('./executions');
@@ -11,7 +10,6 @@ const Execution = require('./executions');
 module.exports = {
   AccessToken,
   AsyncOperation,
-  Granule,
   ReconciliationReport,
   Rule,
   Manager,