+++ packages/api/tests/lambdas/test-create-reconciliation-report.js
@@ -361,7 +361,6 @@ test.before(async (t) => {
 });
 
 test.beforeEach(async (t) => {
-  process.env.GranulesTable = randomId('granulesTable');
   process.env.ReconciliationReportsTable = randomId('reconciliationTable');
 
   t.context.bucketsToCleanup = [];
@@ -372,7 +371,6 @@ test.beforeEach(async (t) => {
   await awsServices.s3().createBucket({ Bucket: t.context.systemBucket })
     .then(() => t.context.bucketsToCleanup.push(t.context.systemBucket));
 
-  await new models.Granule().createTable();
   await new models.ReconciliationReport().createTable();
 
   const cmrSearchStub = sinon.stub(CMR.prototype, 'searchConcept');
@@ -401,7 +399,6 @@ test.afterEach.always(async (t) => {
   await Promise.all(
     flatten([
       t.context.bucketsToCleanup.map(recursivelyDeleteS3Bucket),
-      new models.Granule().deleteTable(),
       new models.ReconciliationReport().deleteTable(),
     ])
   );
@@ -2071,8 +2068,6 @@ test.serial('Internal Reconciliation report JSON is formatted', async (t) => {
   await Promise.all(
     matchingGrans.map((gran) => indexer.indexGranule(esClient, gran, esAlias))
   );
-
-  await new models.Granule().create(matchingGrans);
 
   const event = {
     systemBucket: t.context.systemBucket,