+++ CHANGELOG.md
@@ -78,6 +78,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
     - Remove lambdas/postgres-migration-count-tool, api/endpoints/migrationCounts and api-client/migrationCounts
     - Remove lambdas/data-migration1 tool
     - Remove lambdas/data-migration2 and lambdas/postgres-migration-async-operation
+  - **CUMULUS-2792**
+    - Remove API Granule model and all related test code
+    - Remove granule-csv endpoint
 
 ### Added